Improved colormaps support (#238)
* wip colormaps

* Implement colormap for Image

* Add examples for image colormapping

* Implement colormap for volume and volume slice.

* Add validation examples

* remove clim from mesh material

* also test the new validation example

* refactor image and volume colormap to reuse code

* Implement colormap for Point and Line, and per-vertex colors for mesh.

* fix example

* colors docs

* new black

* flake

+++ examples/multi_slice2.py
@@ -23,7 +23,7 @@ scene.add(background)
 
 scene.add(gfx.AxesHelper(length=50))
 
-vol = imageio.volread("imageio:stent.npz")
+vol = imageio.volread("imageio:stent.npz").astype("float32")
 tex = gfx.Texture(vol, dim=3)
 
 surface = marching_cubes(vol[0:], 200)
